yangilandi

+++ main.py
@@ -7,7 +7,7 @@ from aiogram.utils import executor
 from button.admin_button import main_rp, food_delete, food
 
 from db import Session, MainMenu, Menu
-from funksiyalar.funksiya1 import get_selected_food_name
+from funksiyalar.funksiya1 import get_selected_food_name, get_selected_food_name2
 
 from state.state_uz import Forms, Form
 
@@ -156,15 +156,20 @@ from aiogram.dispatcher import FSMContext
 async def process_delete(query: types.CallbackQuery, state: FSMContext):
     await query.answer()
     selected_food_name = get_selected_food_name()
+    selected_food_name2 = get_selected_food_name2()
     async with state.proxy() as data:
         data['food_name'] = selected_food_name
+        data['food_name'] = selected_food_name2
 
-        if selected_food_name:
+        if selected_food_name and selected_food_name2:
             db = Session()
             food_item = db.query(MainMenu).filter_by(name=selected_food_name).first()
-            if food_item:
+            food_item2 = db.query(Menu).filter_by(name=selected_food_name).first()
+            if food_item and food_item2:
                 food_item.name = selected_food_name
+                food_item2.name = selected_food_name2
                 db.delete(food_item)
+                db.delete(food_item2)
                 db.commit()
             db.close()
             chat_id = query.message.chat.id
